chore: migrate definition methods to own folder
