+++ src/definitions/definitions.ts
@@ -1,0 +1,3 @@
+export * from './joinAllDefs'
+export * from './registerFromCall'
+export * from './registerFromParse'
