+++ src/steps.ts
@@ -7,20 +7,16 @@ import type {
   TStepDefs,
 } from './types'
 
+import { EStepType } from './types'
 import { matcher } from './matcher'
 import { constants } from './constants'
-import { uuid } from '@keg-hub/jsutils'
+import { validateDefinition } from './utils/helpers'
 import { throwNoMatchingStep } from './utils/errors'
-import { EStepMethodType, EStepType } from './types'
-import { sanitize, validateDefinition } from './utils/helpers'
-import { isArr, capitalize, eitherArr, isStr, noOpObj, ensureArr } from '@keg-hub/jsutils'
-import {
-  resolveModule,
-  resolveRequire,
-  resolveGlobalObj,
-} from './utils/globalScope'
+import { isArr, capitalize, isStr, ensureArr } from '@keg-hub/jsutils'
+import { joinAllDefs, registerFromParse, registerFromCall } from './definitions'
 
-const { REGEX_VARIANT, EXPRESSION_VARIANT, STEP_TYPES } = constants
+
+const { STEP_TYPES } = constants
 
 type TRegisterArgs = [
   string|TStepDef[]|TStepDef,
@@ -29,160 +25,6 @@ type TRegisterArgs = [
   TStepDefMethod,
   TStepMeta|undefined
 ]
-type TTempRegisterArgs = [
-  match:string,
-  method:TStepDefMethod,
-  meta:TStepMeta
-]
-
-type TTempContainer = Record<EStepMethodType, TStepDef[]>
-
-/**
- * Builds the text content of a step definition call
- * @function
- * @private
- *
- */
-const getContent = (def:TStepDef) => {
-  const match =
-    def.variant === REGEX_VARIANT ? def.match.toString() : `"${def.match}"`
-
-  return `${capitalize(def.type)}(${match}, ${def.method.toString()})`
-}
-
-const stringToRegex = (str:string) => {
-  const main = str.match(/\/(.+)\/.*/)[1]
-  const options = str.match(/\/.+\/(.*)/)[1]
-
-  return new RegExp(main, options)
-}
-
-/**
- * Registers a step definition by type
- * @function
- * @private
- *
- */
-const registerFromCall = function (
-  internalType:EStepType,
-  type:EStepType,
-  match:string,
-  method:TStepDefMethod,
-  meta:TStepMeta = noOpObj
-) {
-  
-  const variant = match.toString().indexOf('/') === 0 ? REGEX_VARIANT : EXPRESSION_VARIANT
-  const formattedMatch = variant === REGEX_VARIANT
-    ? stringToRegex(match.toString())
-    : match.toString()
-
-  const definition:Partial<TStepDef> = {
-    type,
-    meta,
-    method,
-    variant,
-    // TODO: add token parsing
-    tokens: [],
-    match: formattedMatch,
-  }
-
-  definition.name = sanitize(definition as TStepDef)
-  definition.content = getContent(definition as TStepDef)
-  definition.uuid = uuid()
-
-  const definitions = this.list()
-  const newDefinition = validateDefinition(definition as TStepDef, definitions)
-
-  newDefinition && this[internalType].push(newDefinition)
-
-  return newDefinition
-}
-
-/**
- * Helper method to wrap the default register method of a step definition
- * Allows capturing the definition when it's registered
- * Contains only newly registered definitions, NOT all definitions
- * @function
- * @private
- *
- */
-const tempRegister = (
-  parent:Steps,
-  type:EStepMethodType,
-  container:TTempContainer
-) => {
-  /**
-   * Captures a registered step definition and adds it the the container object
-   * @function
-   * @internal
-   *
-   */
-  return (...args:TTempRegisterArgs) => {
-    const definition = parent[type](...args)
-    container[type].push(definition)
-
-    return definition
-  }
-}
-
-/**
- * Registers a parsed step definition object
- * @function
- * @private
- * @param {Array|string} definitions - Array of strings or single string
- *                                     of the text content form a definition file
- *
- * @returns {void}
- */
-const registerFromParse = function (definitions:string|string[]) {
-  // TRegisterArgs
-  // Ensures a consistent index due to being an array
-  const DEF_TYPES:EStepMethodType[] = this.types.map((type:EStepType) => capitalize(type))
-
-  // Build a container for holding the newly added definitions
-  // Looks like the object below
-  // { Given: [], When: [], Then: [], But: [], And: [] }
-  const container = DEF_TYPES.reduce((built:TTempContainer, type) => {
-    built[type] = []
-    return built
-  }, {} as TTempContainer)
-
-  // Loop over the passed in definitions
-  eitherArr<string[]>(definitions, [definitions]).map(definition => {
-    // Create a dynamic function calling the definition
-    // The definition should be a call to a global Given, When, Then methods
-    // Which is comes from the tempRegister method for each type
-    Function(`return (global, require, module, ${DEF_TYPES.join(',')}) => {
-          return (function(global) { ${definition} }).call(global, global)
-        }`)()(
-      // Pass in the global object so we can bind the dynamic function to it
-      // Allows referencing values on the global scope directly
-      // For example myGlobalFunction() instead of window.myGlobalFunction()
-      resolveGlobalObj(),
-      resolveRequire(),
-      resolveModule(),
-      // Call the tempRegister for each type,
-      // Then spread the response as arguments to the dynamic function
-      ...DEF_TYPES.map(type => tempRegister(this, type, container))
-    )
-  })
-
-  // Return the container which should now hold all newly registered definitions ONLY
-  return container
-}
-
-/**
- * Join all step types together into a single array
- * @function
- * @private
- *
- */
-const joinAllSteps = (instance:Steps):TStepDef[] => {
-  return instance.types.reduce(
-    (stepDefs, type:EStepType) => stepDefs.concat(instance[`_${type}`]),
-    []
-  )
-}
 
 /**
  * Allows registering step definition matchs and functions
@@ -235,7 +77,7 @@ export class Steps {
    *
    */
   list = () => {
-    return joinAllSteps(this)
+    return joinAllDefs(this)
   }
 
   /**
