+++ src/types/index.ts
@@ -7,5 +7,6 @@ export * from './test.types'
 export * from './steps.types'
 export * from './assemble.types'
 export * from './matcher.types'
+export * from './definitions.types'
 export * from './paramTypes.types'
 export * from './indexed.types'
